Problem with tinyint for some drivers

+++ src/main/sql/UNS2-create.sql
@@ -98,7 +98,7 @@ DROP TABLE IF EXISTS `CHANNEL_METADATA_ELEMENTS`;
 CREATE TABLE `CHANNEL_METADATA_ELEMENTS` (
   `CHANNEL_ID` int(10) unsigned NOT NULL default '0',
   `METADATA_ELEMENT_ID` int(10) unsigned NOT NULL default '0',
-  `VISIBLE` tinyint(1) NOT NULL default '0',
+  `VISIBLE` int(10) unsigned NOT NULL default '0',
   `APPEARANCE_ORDER` tinyint(3) unsigned default '0',
   `FILTERED` tinyint(1) NOT NULL default '0',
   `OBSOLETE` tinyint(1) NOT NULL default '0',
